requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
matplotlib
seaborn
scikit-learn